refactor: move route to /events/{event_id}/alerts
